% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling_misc.R
\name{check_features_match_glm}
\alias{check_features_match_glm}
\title{Check whether data features match the ones used in the model}
\usage{
check_features_match_glm(x, glm_model, lambda = "lambda.1se")
}
\arguments{
\item{x}{\code{data.matrix} with data to be scored}

\item{glm_model}{GLM model object}

\item{s}{Optional, lambda parameter (`s` in the \code{glmnet} package)}
}
\description{
Check whether data features match the ones used in the model
}
