% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{check_R_version}
\alias{check_R_version}
\title{Checks R version agains desired}
\usage{
check_R_version(compare_to = NULL, comparison = "equal")
}
\arguments{
\item{comparison}{}
}
\description{
Checks R version agains desired
}
