% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{clean_memory}
\alias{clean_memory}
\title{Clean memory by running \code{gc()} n times.}
\usage{
clean_memory(n = 10)
}
\arguments{
\item{n}{Defaults to 10.}
}
\description{
Clean memory by running \code{gc()} n times.
}
