% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.table_misc.R
\name{copy_from_clipboard}
\alias{copy_from_clipboard}
\title{Copies data from clipboard (e.g. Excel)}
\usage{
copy_from_clipboard(header = TRUE)
}
\arguments{
\item{header}{\code{Logical}, indicates if the data have a header row}
}
\value{
\code{data.table} with pasted data
}
\description{
Copies data from clipboard (e.g. Excel)
}

