% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.table_misc.R
\name{count_NAs}
\alias{count_NAs}
\title{Print sum(s) of missing values}
\usage{
count_NAs(data_in, prop = FALSE, print = TRUE)
}
\arguments{
\item{prop}{Flag indicating if results should include proportions.}

\item{print}{Flag indicating if results should be printed.}

\item{data}{Input object.}
}
\value{
Table with missing values counts (and proportions).
}
\description{
Prints sum(s) of missing values (coded as \code{NA}) in an object
(\code{vector}, \code{data.frame}, \code{data.table} etc.).
}
\examples{
dt <- data.table::data.table(a = rnorm(n=1000, mean=20, sd=5),
                             b = rnorm(n=1000, mean=20, sd=5),
                             c = rnorm(n=1000, mean=20, sd=5))
dt[sample.int(nrow(dt), round(nrow(dt) * 0.15)), a := NA_real_]
count_NAs(dt)
count_NAs(dt, prop = TRUE)
count_NAs(dt[, a])
count_NAs(dt[, a], prop = TRUE)
count_NAs(dt[, .(a)])
count_NAs(dt[, .(a)], prop = TRUE)
}

