% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.table_misc.R
\name{count_values}
\alias{count_values}
\title{Prints count(s) of specified values}
\usage{
count_values(dt_in, value = "", prop = FALSE, na.rm = TRUE,
  print = TRUE)
}
\arguments{
\item{print}{}
}
\value{
Table with counts (and proportions if TRUE) of specified values.
}
\description{
Prints count(s) of specified values in a \code{data.table} object.
}
\examples{
dt <- data.table::data.table(a = rnorm(n=1000, mean=20, sd=5),
                             b = rnorm(n=1000, mean=20, sd=5),
                             c = rnorm(n=1000, mean=20, sd=5))
dt[sample.int(nrow(dt), round(nrow(dt) * 0.15)), a := -Inf]
count_values(dt)
count_values(dt, prop = TRUE)
count_values(dt, -Inf)
count_values(dt, -Inf, prop = TRUE)

TODO: Write tests
}
