% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling_misc.R
\name{cr_table}
\alias{cr_table}
\title{Calculates conversion ratio table with counts and proportions}
\usage{
cr_table(dt, cols, target, order = "cols", prop_count = TRUE)
}
\arguments{
\item{order}{}
}
\description{
Calculates conversion ratio table with counts and proportions
}
\examples{
n <- 200
sample_dt <- data.table::data.table(
  ColA = sample(letters[1:5], n, replace = TRUE),
  ColB = sample(letters[1:5], n, replace = TRUE),
  Target = runif(n) <= 0.1)

cr_table(sample_dt, 'ColA', 'Target', prop_count = FALSE)
cr_table(sample_dt, 'ColA', 'Target', prop_count = TRUE)
cr_table(sample_dt, 'ColA', 'Target')

cr_table(sample_dt, c('ColA','ColB'), 'Target')
cr_table(sample_dt, c('ColA','ColB'), 'Target', order = 'cr')
cr_table(sample_dt, c('ColA','ColB'), 'Target', order = 'count')
cr_table(sample_dt, c('ColA','ColB'), 'Target', order = 'prop')
}
