% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{cramersV_mat}
\alias{cramersV_mat}
\title{Calculater Cramer's V matrix for categorical data}
\usage{
cramersV_mat(dt)
}
\arguments{
\item{levelsMax}{}
}
\description{
Calculater Cramer's V matrix for categorical data
}
