% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.table_misc.R
\name{drop_cols}
\alias{drop_cols}
\title{Drop columns in \code{data.table} by reference or returning a copy.}
\usage{
drop_cols(dt, cols, modify = FALSE)
}
\arguments{
\item{dt}{\code{data.table} to process}

\item{cols}{Column names to drop, vector of strings}

\item{modify}{Boolean, indicating whether columns should be dropped by reference of a copy should be returned}
}
\description{
Drop columns in \code{data.table} by reference or returning a copy.
}
\examples{
require(data.table)
dt <- data.table(V1 = 1, V2 = 2, V3 = 3)
print(dt)
dt_dropped <- drop_cols(dt, c("V1", "V2"), modify = FALSE)
print(dt)
print(dt_dropped)
drop_cols(dt, c("V1", "V2"), modify = TRUE)
print(dt)
}
