% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{dump_to_csv}
\alias{dump_to_csv}
\title{Save data set to \code{dump_dir} folder}
\usage{
dump_to_csv(dt, file_name = paste0(deparse(substitute(dt))),
  dump_dir = "C:/Users/dziedzm/Desktop/temp_file_dump")
}
\arguments{
\item{dt}{Data set to save}

\item{file_name}{File name}

\item{dump_dir}{Folder}
}
\description{
Save data set to \code{dump_dir} folder
}
