% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_clean_and_move_somewhere_else.R
\name{extract_cat_cols}
\alias{extract_cat_cols}
\title{Title}
\usage{
extract_cat_cols(dt, levels_max = NULL, levels_min = 1, exclude = NA)
}
\arguments{
\item{levels_max}{}
}
\description{
Title
}
\examples{
library(data.table)
dt <- data.table(
  V1 = rep(1.1, 10),
  V2 = rep(2L, 10),
  V3 = rep('a',10),
  V4 = letters[1:10])
extract_cat_cols(dt, levels_max = NULL, levels_min = 1)
extract_cat_cols(dt, levels_max = 5, levels_min = 1)
extract_cat_cols(dt, levels_max = NULL, levels_min = 5)
extract_cat_cols(dt, exclude = 'V3')

# Args check:
extract_cat_cols(dt, levels_max = -1, levels_min = 1)
extract_cat_cols(dt, levels_max = NULL, levels_min = -1)
extract_cat_cols(dt, levels_max = 1, levels_min = 5)
dt
}
