% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_clean_and_move_somewhere_else.R
\name{extract_cont_cols}
\alias{extract_cont_cols}
\title{Title}
\usage{
extract_cont_cols(dt, exclude = NA)
}
\arguments{
\item{dt}{}
}
\description{
Title
}
\examples{
library(data.table)
dt <- data.table(V1 = 1.1, V2 = 2L, V3 = 'a')
str(dt)
extract_cont_cols(dt)
extract_cont_cols(dt, exclude = 'V1')
extract_cont_cols(dt, exclude = 'V3')
dt
}
