% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{file_rename_with_dir_create}
\alias{file_rename_with_dir_create}
\title{Move file and create dir if needed}
\usage{
file_rename_with_dir_create(from, to)
}
\arguments{
\item{from}{Origin file path}

\item{to}{Destination file path}
}
\description{
http://stackoverflow.com/a/10268255/4272484
}
