% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.table_misc.R
\name{fill_NAs_with_const}
\alias{fill_NAs_with_const}
\title{Fills missing values in data.table with a specified constant value.}
\usage{
fill_NAs_with_const(data_in, const_val = 0, cols_to_fill = names(data_in))
}
\arguments{
\item{cols_to_fill}{}
}
\description{
Fills missing values in data.table with a specified constant value.
}
