% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.table_misc.R
\name{fill_values_with_NA}
\alias{fill_values_with_NA}
\title{Fills specified values (def. empty strings) in data.table with an NA.}
\usage{
fill_values_with_NA(data_in, values_to_fill = "",
  cols_to_fill = names(data_in))
}
\arguments{
\item{cols_to_fill}{}
}
\description{
Fills specified values (def. empty strings) in data.table with an NA.
}
