% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv_related.R
\name{fread_CSVs_from_dir}
\alias{fread_CSVs_from_dir}
\title{Read (using \code{data.table::fread()}) and rbind files from directory}
\usage{
fread_CSVs_from_dir(dir_, verbose = FALSE, fill = TRUE, pattern = NULL)
}
\arguments{
\item{dir_}{Directory path}

\item{verbose}{Boolean switch for printing file names etc.}

\item{fill}{If \code{TRUE} fills missing columns with NAs. By default \code{FALSE}.}

\item{pattern}{Character string to filter files}
}
\description{
Read (using \code{data.table::fread()}) and rbind files from directory
}
