% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv_related.R
\name{fread_CSVs_from_list}
\alias{fread_CSVs_from_list}
\title{Read (using \code{data.table::fread()}) and rbind files from list}
\usage{
fread_CSVs_from_list(file_list, verbose = FALSE, fill = TRUE,
  pattern = NULL)
}
\arguments{
\item{file_list}{Character vector from e.g. \code{list.files()} function}

\item{verbose}{Boolean switch for printing file names etc.}

\item{fill}{If \code{TRUE} fills missing columns with NAs. By default \code{FALSE}.}
}
\description{
Read (using \code{data.table::fread()}) and rbind files from list
}
