% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv_related.R
\name{fread_most_recent_file}
\alias{fread_most_recent_file}
\title{Read (using \code{data.table::fread()}) most recent file from directory}
\usage{
fread_most_recent_file(dir_, pattern = NULL, oldest = FALSE)
}
\arguments{
\item{dir_}{Directory path}

\item{pattern}{Character string to filter files}

\item{oldest}{Boolean switch to reverse ordering}
}
\description{
Read (using \code{data.table::fread()}) most recent file from directory
}
