% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_active_file_path}
\alias{get_active_file_path}
\title{Funtion returns the active file's path}
\usage{
get_active_file_path()
}
\value{
Active file's path
}
\description{
Funtion returns the active file's path
}
\examples{
getActiveFilePath()
}

