% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_running_script_dir_cmd}
\alias{get_running_script_dir_cmd}
\title{Returns normalized dir path to running script}
\usage{
get_running_script_dir_cmd()
}
\description{
Returns normalized dir path to running script
}
