% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{getActiveFilePath}
\alias{getActiveFilePath}
\title{Funtion returns the active file's path}
\usage{
getActiveFilePath()
}
\value{
Active file's path
}
\description{
Funtion returns the active file's path
}
\examples{
getActiveFilePath()
}

