% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{gg_color_palette}
\alias{gg_color_palette}
\title{Recreate ggplot2 colour palette}
\usage{
gg_color_palette(n)
}
\arguments{
\item{n}{Number of colours}
}
\value{
\code{Vector} of colours
}
\description{
Recreate ggplot2 colour palette
}
\examples{
gg_color_palette(3)
}

