% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling_misc.R
\name{gini}
\alias{gini}
\title{Gini score}
\usage{
gini(actual, predictions)
}
\arguments{
\item{actual}{Actual outcome}

\item{predictions}{Predicted outcome}
}
\value{
Gini score
}
\description{
Gini score for binary outcome modelling
(rather use normalized version: mdmisc::gini_norm())
}
\details{
Based on \link{https://www.kaggle.com/wiki/RCodeForGini}
}
