% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling_misc.R
\name{gini_norm}
\alias{gini_norm}
\title{Normalized Gini score}
\usage{
gini_norm(actual, predictions)
}
\arguments{
\item{actual}{Actual outcome}

\item{predictions}{Predicted outcome}
}
\value{
Normalized Gini score
}
\description{
Normalized Gini score for binary outcome modelling
}
\details{
Based on \link{https://www.kaggle.com/wiki/RCodeForGini}
}
\examples{
df <- data.frame(
  actual = c(1,1,1,1,1,0,1,1,1,0,1,1,0,1,1,1,1,0,0,0,1,0,1,0,0,0,1,0,0,0,0,0,0,1,0,0,0,0,0,0),
  predicted = seq(1, 0, length.out = 40)
)
gini_val <- gini_norm(df[, 'actual'], df[, 'predicted'])
## Compare to AUC from pROC package
# AUC_val <- pROC::roc(df[, 'actual'], df[, 'predicted'])$auc
# gini_val == 2 * AUC_val - 1
}
