% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouping_and_binning.R
\name{group_by_threshold}
\alias{group_by_threshold}
\title{Group categories by (cumulative) frequencies}
\usage{
group_by_threshold(dt, feature, threshold = NULL, cum_threshold = NULL,
  no_of_categories = NULL, return_data = FALSE, modify = FALSE,
  other_cat_name = "OTHER")
}
\arguments{
\item{other_cat_name}{}
}
\description{
Group categories by (cumulative) frequencies
}
\examples{
library(mdmisc)
set.seed(2016)
dt <- data.table::data.table('cat' = sample(letters[1:10], 20, replace = TRUE), 'cont' = rnorm(20))

## View cumulative frequencies
group_by_threshold(dt, 'cat')
group_by_threshold(dt, 'cat', threshold = 0.10)
group_by_threshold(dt, 'cat', cum_threshold = 0.90)
group_by_threshold(dt, 'cat', no_of_categories = 3)

## Group categories below 10\\\% of frequency
dt_mod <- group_by_threshold(dt, 'cat', threshold = 0.10, return_data = TRUE)
group_by_threshold(dt_mod, 'cat')

## Group bottom 10\\\% categories based on cumulative frequency
dt_mod <- group_by_threshold(dt, 'cat', cum_threshold = 0.90, return_data = TRUE)
group_by_threshold(dt_mod, 'cat')

## Leave 3 categories based on frequency
dt_mod <- group_by_threshold(dt, 'cat', no_of_categories = 3, return_data = TRUE)
group_by_threshold(dt_mod, 'cat')

## Group and modify in place
group_by_threshold(dt, 'cat', threshold = 0.1, return_data = TRUE, modify = TRUE)
group_by_threshold(dt, 'cat')
}
