% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipak.R
\name{ipak}
\alias{ipak}
\title{Unified installer for CRAN, Bioconductor, and GitHub packages}
\usage{
ipak(pkg, repository = c("CRAN", "Bioconductor", "github"))
}
\arguments{
\item{pkg}{package names in a character vector}

\item{repository}{one of "CRAN", "Bioconductor", "github".
Indicates from which repository to install packages.}
}
\description{
Takes a vector of package names and installs them.
Requires specifying repository type.
}
\note{
Modified after \href{https://bvieth.github.io/powsimR/}{powsimR installation instructions}
}
\examples{
\dontrun{
# CRAN packages
cranpackages <- c("tidyverse", "data.table")
ipak(cranpackages, repository = "CRAN")
# Bioconductor
biocpackages <- c("limma", "SingleCellExperiment")
ipak(biocpackages, repository = "Bioconductor")
# GitHub
githubpackages <- c("nanxstats/ggsci", "slowkow/ggrepel")
ipak(githubpackages, repository = "github")
}
}
