% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{is.not.na}
\alias{is.not.na}
\title{Checks for an argument not being NA}
\usage{
is.not.na(x)
}
\arguments{
\item{x}{}
}
\value{
\code{Boolean} indicating whether x is not NA
}
\description{
Checks for an argument not being NA
}
