% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{is.not.null}
\alias{is.not.null}
\title{Checks for an argument not being NULL}
\usage{
is.not.null(x)
}
\arguments{
\item{x}{Input argument}
}
\value{
\code{Boolean} indicating whether x is not NULL
}
\description{
Checks for an argument not being NULL
}
