% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.table_misc.R
\name{leave_cols}
\alias{leave_cols}
\title{Leave selected columns in input \code{data.table}}
\usage{
leave_cols(dt, col_names, modify = FALSE, print = TRUE)
}
\arguments{
\item{dt}{Input \code{data.table}}

\item{col_names}{Column names to leave}

\item{modify}{Logical, whether to modify dt or return copy}
}
\value{
Modified dt (if \code{modify == TRUE}) or copy of dt with removed columns
}
\description{
Leave selected columns in input \code{data.table}
}
