% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{lib_pacman}
\alias{lib_pacman}
\title{Simple wrapper for \code{pacman::p_load}}
\usage{
lib_pacman(package)
}
\arguments{
\item{package}{}
}
\value{
None (invisible NULL).
}
\description{
Function name matches the first characters of \code{library}
}
\examples{
lib_pacman(ggplot2)
lib_pacman('ggplot2')
}
