% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{list_csv_files}
\alias{list_csv_files}
\title{Lists CSV files in the specified directory}
\usage{
list_csv_files(path = ".", full_names = FALSE)
}
\arguments{
\item{full_names}{}
}
\value{
List of CSV files in the directory
}
\description{
Lists CSV files in the specified directory
}
