% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{mark_outliers_DBSCAN}
\alias{mark_outliers_DBSCAN}
\title{Find outliers using density based method}
\usage{
mark_outliers_DBSCAN(x, na.rm = TRUE, quant = 0.15, dist_method = "euc",
  ...)
}
\arguments{
\item{...}{}
}
\description{
Function returns \code{logical} \code{vector} indicating whether a value lies
outside of density based clusters
}

