% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{mark_outliers_IQR}
\alias{mark_outliers_IQR}
\title{Find outliers using 1.5 IQR method}
\usage{
mark_outliers_IQR(x, na.rm = TRUE, quants = c(0.25, 0.75), ...)
}
\arguments{
\item{x}{Processed \code{vector}}

\item{na.rm}{Remove NAs from \code{x}}

\item{quants}{Which quantiles should be used, defaults to \code{c(.25, .75)}}

\item{...}{Other parameters passed to \code{quantile()}}
}
\value{
\code{logical} \code{vector} indicating whether a value is an outlier
}
\description{
Function returns \code{logical} \code{vector} indicating whether a value lies
within 1.5 * \code{IQR} from specified quantiles of the \code{numeric}
\code{vector}
}
\examples{
mark_outliers_IQR(rnorm(20, 0, 1))
}

