% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{mark_outliers_quant}
\alias{mark_outliers_quant}
\title{Find outliers using 0.01 and 0.99 quantiles}
\usage{
mark_outliers_quant(x, na.rm = TRUE, quants = c(0.01, 0.99),
  return_params = FALSE, ...)
}
\arguments{
\item{x}{Processed \code{vector}}

\item{na.rm}{Remove NAs from \code{x}}

\item{quants}{Which quantiles should be used, defaults to \code{c(0.01, 0.99)}}

\item{...}{Other parameters passed to \code{quantile()}}
}
\value{
\code{logical} \code{vector} indicating whether a value is an outlier
}
\description{
Function returns \code{logical} \code{vector} indicating whether a value lies
outside of the 0.01 and 0.99 quantiles.
}
\examples{
mark_outliers_quant(rnorm(20, 0, 1))
}
