% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.table_misc.R
\name{move_columns}
\alias{move_columns}
\title{Applies \code{move_columns} to a \code{data.table}.}
\usage{
move_columns(data_in, move_command)
}
\arguments{
\item{data_in}{Input \code{data.table} object.}

\item{move_command}{Rearrange command string.}
}
\value{
Rearranged input object.
}
\description{
Applies \code{move_columns} to a \code{data.table}.
}
\examples{
dt <- data.table::data.table(a = 1:3, b = 1:3, g = 1:3)
move_columns(dt, "g first")
move_columns(dt, "g first; a last")
}

