% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_binary_outcome}
\alias{plot_binary_outcome}
\title{Plots binary outcome}
\usage{
plot_binary_outcome(df, pred_name, actual_name, x_lab = NULL, y_lab = NULL,
  theme = NULL, type = "pipes", smooth = TRUE, smoothing_level = NA)
}
\arguments{
\item{df}{Data}

\item{pred_name}{Predictions' column name}

\item{actual_name}{Actual values' column name}

\item{x_lab}{X-axis label}

\item{y_lab}{Y-axis label}

\item{theme}{Theme function to use}

\item{type}{Plot type, \code{type = c('pipes', 'jitter')}}

\item{smooth}{Logical, whether plot smoothing line or not}

\item{smoothing_level}{If \code{smooth == TRUE} the smoothing level, otherwise ignored}
}
\value{
Plot object
}
\description{
Plots binary outcome, use for binary models outcomes etc.
}
\examples{
df <- data.frame(
  actual = c(1,1,1,1,1,0,1,1,1,0,1,1,0,1,1,1,1,0,0,0,1,0,1,0,0,0,1,0,0,0,0,0,0,1,0,0,0,0,0,0),
  predicted = seq(1, 0, length.out = 40)
)
plot_binary_outcome(df, 'predicted', 'actual')
plot_binary_outcome(df, 'predicted', 'actual', smoothing_level = 0.75)
plot_binary_outcome(df, 'predicted', 'actual', x_lab = 'Predicted', y_lab = 'Actual')
plot_binary_outcome(df, 'predicted', 'actual', theme = theme_bw())
plot_binary_outcome(df, 'predicted', 'actual', type = 'pipes', theme = theme_bw())
plot_binary_outcome(df, 'predicted', 'actual', type = 'jitter', theme = theme_bw())
}
