% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_clean_and_move_somewhere_else.R
\name{plot_cor_cont}
\alias{plot_cor_cont}
\title{Plots correlation matrix for continuous variables}
\usage{
plot_cor_cont(dt, print = TRUE)
}
\arguments{
\item{print}{}
}
\description{
Plots correlation matrix for continuous variables
}
\examples{
library(data.table); library(magrittr)
data(mtcars)
mtcars \%>\% data.table() \%>\% plot_cor_cont()
}
