% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_lift_chart}
\alias{plot_lift_chart}
\title{Plots lift chart using binary target and model predictions}
\usage{
plot_lift_chart(target, predictions, breaks_no = 4,
  xlabel = "Predicted response", ylabel = "Actual conversion ratio",
  title = "Lift chart", file_name = "lift_chart", withCounts = FALSE,
  showPlot = TRUE, saveToFile = FALSE, returnPlot = FALSE)
}
\arguments{
\item{returnPlot}{}
}
\description{
Plots lift chart using binary target and model predictions
}
