% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{read_pass}
\alias{read_pass}
\title{Read password from specified file}
\usage{
read_pass(file_path)
}
\arguments{
\item{file}{Text file with password stored}
}
\value{
String with password
}
\description{
Read password from specified file
}

