% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{remove_single_value_cols}
\alias{remove_single_value_cols}
\title{Title}
\usage{
remove_single_value_cols(dt, exclude_list = NULL, verbose = FALSE)
}
\arguments{
\item{exclude_list}{}
}
\description{
Title
}
\examples{
require(data.table)
dt <- data.table(V1 = 1, V2 = 2, V3 = 3)
dt
remove_single_value_cols(dt)
}
