% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.table_misc.R
\name{reorder_col}
\alias{reorder_col}
\title{Applies \code{reorder_vec} to a \code{data.table}.}
\usage{
reorder_col(data_in, reorder_command)
}
\arguments{
\item{data_in}{Input \code{data.table} object.}

\item{reorder_command}{Rearrange command string.}
}
\value{
Rearranged input object.
}
\description{
Applies \code{reorder_vec} to a \code{data.table}.
}
\examples{
dt <- data.table::data.table(a = 1:3, b = 1:3, g = 1:3)
reorder_col(dt, "g first")
reorder_col(dt, "g first; a last")
}

