% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.table_misc.R
\name{reorder_vec}
\alias{reorder_vec}
\title{Rearrange character vector.}
\usage{
reorder_vec(vec_in, reorder_command)
}
\arguments{
\item{vec_in}{Input character vector.}

\item{reorder_command}{Rearrange command.}
}
\value{
Rearranged \code{vec_in}.
}
\description{
\code{reorder_vec} rearranges and returns an input character vector.
Possible "rearrange" words: c("before", "after", "first", "last")
}
\details{
Source: http://stackoverflow.com/a/18540144/4272484 with minor amendments
}
\examples{
reorder_vec(c("a", "b", "g"), "g first")
df <- data.frame(a = 1:3, b = 1:3, g = 1:3)
reorder_vec(names(df), "g first")
reorder_vec(names(df), "g first; a last")
}

