% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{rm_all_exc_func}
\alias{rm_all_exc_func}
\title{Remove all user defined variables except functions}
\usage{
rm_all_exc_func(ask = TRUE)
}
\arguments{
\item{ask}{\code{logical} indicating whether user confirmation is needed.}
}
\value{
None (invisible NULL).
}
\description{
Remove all user defined variables except functions
}
\examples{
rm_all_exc_func()  # Will ask for confirmation.
}
