% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.table_misc.R
\name{sample_and_replace_value}
\alias{sample_and_replace_value}
\title{Replace given value in a random pct of rows (by grouping)}
\usage{
sample_and_replace_value(dt, pct, col_name, value, grouping = NULL,
  sort = FALSE)
}
\arguments{
\item{grouping}{}
}
\description{
Function modifies the input dt!
}
\examples{
set.seed(2016)
size <- 10
dt <- data.table(
  id = 1:size, A = sample(letters[1:3], size, replace = TRUE), B = 'N',
  C = sample(1:100, size, replace = TRUE) + sample(30:70, size, replace = TRUE))
set.seed(2016)
sample_and_replace_value(dt, pct = 0.5, col_name = 'B', value = 'Y', grouping = 'A')
dt[, table(A, B)]
sample_and_replace_value(dt, pct = 0.5, col_name = 'B', value = 'Y', grouping = 'A', sort = TRUE)
dt
}
