% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.table_misc.R
\name{sample_ids}
\alias{sample_ids}
\title{Return a random sample of row ids (by grouping)}
\usage{
sample_ids(dt, pct, grouping = NULL, sort = FALSE,
  return_grouping = FALSE)
}
\arguments{
\item{sort}{}
}
\description{
Return a random sample of row ids (by grouping)
}
\examples{
set.seed(2016)
size <- 10
dt <- data.table(
  id = 1:size, A = sample(letters[1:3], size, replace = TRUE), B = 'N',
  C = sample(1:100, size, replace = TRUE) + sample(30:70, size, replace = TRUE))
set.seed(2016)
sample_ids(dt, 0.5, grouping = 'A')
set.seed(2016)
sample_ids(dt, 0.5, grouping = 'A', sort = TRUE)
set.seed(2016)
sample_ids(dt, 0.5, grouping = 'A', return_grouping = TRUE)
set.seed(2016)
sample_ids(dt, 0.5, grouping = 'A', sort = TRUE, return_grouping = TRUE)
}
