% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.table_misc.R
\name{sample_rows}
\alias{sample_rows}
\title{Title}
\usage{
sample_rows(dt, pct, grouping, sort = FALSE)
}
\arguments{
\item{dt}{\code{data.table} to sample from}

\item{pct}{Percentage of data sampled}

\item{grouping}{Grouping (a \code{character vector})}

\item{sort}{c('no', 'id', 'gr')}
}
\value{
\code{data.table} with sampled data
}
\description{
Title
}
\examples{
set.seed(2016)
size <- 10
dt <- data.table(
  id = 1:size, A = sample(letters[1:3], size, replace = TRUE), B = 'N',
  C = sample(1:100, size, replace = TRUE) + sample(30:70, size, replace = TRUE))
set.seed(2016)
sample_rows(dt, 0.5, grouping = 'A')
sample_rows(dt, 0.5, grouping = 'A', sort = TRUE)
sample_rows(dt, 0.5, grouping = c('A','B'), sort = TRUE)
}
