% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emailing.R
\name{send_test_email_mailR}
\alias{send_test_email_mailR}
\title{Sends test email using \code{send_email_mailR} function}
\usage{
send_test_email_mailR()
}
\value{
See \code{mailR::send.mail()}
}
\description{
Sends test email using \code{send_email_mailR} function
}
\examples{
send_test_email_mailR()
}
