% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{setdiff_two_way}
\alias{setdiff_two_way}
\title{Two way set difference}
\usage{
setdiff_two_way(first, second)
}
\arguments{
\item{first}{Input argument}

\item{second}{Input argument}
}
\description{
Two way set difference
}
