% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{setup_proxy}
\alias{setup_proxy}
\title{Setup proxy settings}
\usage{

  setup_proxy(file_path = "F:/Marketing/Group Marketing/Database Marketing/Marketing Analytics/Personal/Mateusz/00 DandG proxy settings/proxy_settings.csv",
  url = NULL, port = NULL, username = NULL,
  password = getPass::getPass())
}
\arguments{
\item{url}{Proxy url}

\item{port}{Proxy port}

\item{username}{Proxy username}

\item{password}{Proxy password}
}
\value{
\code{TRUE}
}
\description{
Setup proxy settings
}
\examples{
mdmisc::setup_proxy() # Runs with default file_path
}
