% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{single_microbenchmark}
\alias{single_microbenchmark}
\title{\code{microbenchmark::microbenchmark} with default \code{reps = 1}}
\usage{
single_microbenchmark(expr, reps = 1)
}
\arguments{
\item{expr}{Expressions to benchmark.}

\item{reps}{No. of replications.}
}
\value{
Object created and benchmark results.
}
\description{
\code{microbenchmark::microbenchmark} with default \code{reps = 1}
}
\examples{

}
