% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{sorted_names}
\alias{sorted_names}
\title{Returns sorted names of the object}
\usage{
sorted_names(obj)
}
\arguments{
\item{obj}{Object to process}
}
\value{
\code{character} \code{vector} of object's sorted names
}
\description{
Returns sorted names of the object
}
\examples{
df <- data.frame(b = 1, a = 2); names(df); sorted_names(df)
}
