% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_processing.R
\name{str_to_CamelCase}
\alias{str_to_CamelCase}
\title{Convert 'string' to 'String', or 'this.string' to 'ThisString'}
\usage{
str_to_CamelCase(x, split = " ", collapse = "")
}
\arguments{
\item{x}{Input character vector}

\item{split}{Pattern/char(s) to split on}

\item{collapse}{Char(s) to join with}
}
\description{
Based on https://stackoverflow.com/a/11672354/4272484
}
\examples{
t1 <- c('this text', 'next text')
t2 <- c('this.text', 'next.text')
t3 <- 'this'
t4 <- 'this text'

str_to_CamelCase(t1) # c('ThisText', 'NextText')
str_to_CamelCase(t1, collapse = ' ') # c('This Text', 'Next Text')
str_to_CamelCase(t2, split = '.', collapse = ' ') # c('This Text', 'Next Text')

str_to_CamelCase(t3) # 'This'
str_to_CamelCase(t4) # 'ThisText'
}
