% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_processing.R
\name{str_to_Capital}
\alias{str_to_Capital}
\title{Helper function for str_to_CamelCase, convert 'string' to 'String'}
\usage{
str_to_Capital(x)
}
\arguments{
\item{x}{Input string}
}
\value{

}
\description{
Helper function for str_to_CamelCase, convert 'string' to 'String'
}
