% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params_and_templates.R
\name{substitute_params}
\alias{substitute_params}
\title{Title}
\usage{
substitute_params(params_table, template_files,
  target_dir = get_active_file_path(),
  cols_to_ignore = c("PARAM_FILE_NAME_DATE", "COMMENTS"))
}
\arguments{
\item{params_table}{Table (\code{data.frame} or \code{data.table}) with parameters to substitute.}

\item{template_files}{List of template files to substitute the params in.}

\item{target_dir}{Directory where to create files, defaults to active file directory.}

\item{cols_to_ignore}{Columns from the parameters table \code{params_table} to ignore when substituting, defaults to \code{c('PARAM_FILE_NAME_DATE', 'COMMENTS')}.}
}
\description{
Title
}
