% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.table_misc.R
\name{table_data_table}
\alias{table_data_table}
\title{Counts and percentages table by a specified colname(s)}
\usage{
table_data_table(data_in, colname)
}
\arguments{
\item{data_in}{\code{data.table} to process}

\item{colname}{Column name(s) to calculate counts and parcentages}
}
\value{
\code{data.table} with counts and parcentages
}
\description{
Counts and percentages table by a specified colname(s)
}
\examples{
set.seed(2016)
dt <- data.table::data.table(A = sample(LETTERS[1:10], 200, replace = TRUE),
                             B = sample(LETTERS[1:10], 200, replace = TRUE))
table_data_table(dt, "A")
table_data_table(dt, "B")
table_data_table(dt, c("A", "B"))
}
