% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unembed.R
\name{unembed}
\alias{unembed}
\title{Splits rows based on values embedded in a column based on a separator}
\usage{
unembed(data, col, sep, fixed = TRUE)
}
\arguments{
\item{data}{a data.frame object, with a column containing values with multiple
variables.}

\item{col}{column name containing values with multiple variables.}

\item{sep}{character/regular expression to split a column by}

\item{fixed}{logical. If TRUE (default), match split exactly, otherwise use regular expressions.}
}
\value{
a data frame with columns containing unique variables obtained by
unembedding.
}
\description{
Splits rows based on values embedded in a column based on a separator.
E.g. one row with "ABC11 /// BCD22" variable in a column in will be split into
two separate entries (" /// " separator), creating two ABC11 and BCD22 rows
with other values equal to the original row.
}
\note{
Courtesy to \link[https://github.com/aaronwolen]{Aaron Wolen}
}
\examples{
\dontrun{
df <- data.frame(key = c("a", "a;b", "a;b;c"), val = 1:3)
unembed(df, "key", ";")
}
}
