% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{yes_no_input}
\alias{yes_no_input}
\title{Internal function to parse user 'yes/no' input}
\usage{
yes_no_input(message_text = NULL)
}
\value{
User input if matching one of \code{c('y','yes','Yes','n','no','No')}
}
\description{
Internal function to parse user 'yes/no' input
}
