% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{dump_to_csv}
\alias{dump_to_csv}
\title{Save data set to \code{dump_dir} folder}
\usage{
dump_to_csv(dt, file_name = NULL, dump_dir = NULL, ...)
}
\arguments{
\item{dt}{Data to save}

\item{file_name}{File name}

\item{dump_dir}{Folder directory}

\item{...}{Arguments passed to \code{data.table::fwrite} function}
}
\description{
Warning: Conversion to \code{data.frame} tested only from \code{matrix}
}
\examples{
x <- data.frame(A = c(1:3), B = letters[1:3])
dump_to_csv(x)
dump_to_csv(data.frame(A = c(1:3), B = letters[1:3]))
dump_to_csv(data.frame(A = c(1:3), B = letters[1:3]), file_name = 'a.csv')
# Conversion to data.frame:
dump_to_csv(as.matrix(c(1:4)))
# Errors:
dump_to_csv(x, file_name = ':a.csv')
dump_to_csv(data.frame(A = c(1:3), B = letters[1:3]), file_name = ':a.csv')
# Not run: other scenarios not tested!
}
