% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouping_and_binning.R
\name{group_by_threshold}
\alias{group_by_threshold}
\title{Title}
\usage{
group_by_threshold(dt, feature, threshold = NULL, cum_threshold = NULL,
  no_of_categories = NULL, return_data = FALSE, modify = FALSE,
  other_cat_name = "OTHER")
}
\arguments{
\item{other_cat_name}{}
}
\description{
Title
}
\examples{
set.seed(2016)
dt <- data.table::data.table('cat' = sample(letters[1:10], 20, replace = TRUE), 'cont' = rnorm(20))

# View cumulative frequencies
group_by_threshold(dt, 'cat')

# Group bottom 20\\\% categories based on cumulative frequency
dt_mod <- group_by_threshold(dt, 'cat', cum_threshold = 0.2, return_data = TRUE)
dt_mod

# Group and modify in place
group_by_threshold(dt, 'cat', cum_threshold = 0.2, return_data = TRUE, modify = TRUE)
dt
}
