% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{merge_csv}
\alias{merge_csv}
\title{Title}
\usage{
merge_csv(path, str_to_match, suffix = "_merged")
}
\arguments{
\item{str_to_match}{}
}
\description{
Title
}
\examples{
path <- mdmisc::get_active_file_path()
mdmisc::merge_csv(path, 'string to match')
}
