% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{setup_proxy}
\alias{setup_proxy}
\title{Setup proxy settings}
\usage{
setup_proxy(file_path = NA_character_, url = NULL, port = NULL,
  username = NULL, password = getPass::getPass())
}
\arguments{
\item{url}{Proxy url}

\item{port}{Proxy port}

\item{username}{Proxy username}

\item{password}{Proxy password}
}
\value{
\code{TRUE} if successful.
}
\description{
Setup proxy settings
}
\examples{
setup_proxy(url = "1.1.1.1", port = 1111, username = "domain//username")
## Not run: file template
##       url, port,           username
## "1.1.1.1", 1111, "domain//username"
}
